gacha-gen: split per table generation to separate file
